AI-Image-Generator/requirements.txt
# AI Image Generator Requirements
openai==0.28.1
Pillow==10.0.1
requests==2.31.0